lift TextureAtlas out of Draw2d
