+++ src/graphics/draw2d/graphics_pipeline.rs
@@ -1,6 +1,7 @@
-use super::{descriptor_sets, texture_atlas, Vertex};
-use crate::graphics::vulkan::{
-    ffi, shader_module::ShaderModule, Device, Swapchain,
+use super::{descriptor_sets, Vertex};
+use crate::graphics::{
+    texture_atlas::MAX_SUPPORTED_TEXTURES,
+    vulkan::{ffi, shader_module::ShaderModule, Device, Swapchain},
 };
 
 use anyhow::{Context, Result};
@@ -50,9 +51,7 @@ impl GraphicsPipeline {
                 size: std::mem::size_of::<u32>(),
                 ..Default::default()
             }])
-            .data(unsafe {
-                ffi::any_as_u8_slice(&texture_atlas::MAX_SUPPORTED_TEXTURES)
-            })
+            .data(unsafe { ffi::any_as_u8_slice(&MAX_SUPPORTED_TEXTURES) })
             .build();
         let fragment_create_info = vk::PipelineShaderStageCreateInfo::builder()
             .stage(vk::ShaderStageFlags::FRAGMENT)
