+++ src/graphics/texture_atlas/atlas_version.rs
@@ -1,0 +1,9 @@
+use super::AtlasVersion;
+
+impl AtlasVersion {
+    /// A binding revision which will always be considered 'out_of_date'
+    /// relative to the atlas.
+    pub fn out_of_date() -> Self {
+        Self { revision_count: 0 }
+    }
+}
